fix bug with start of file comments (#1)
* fix bug with start of file comments

refactors the comment parsing code outside of the token code such that
token parsing does not restrict the comment parsing code. Prior to this
commit the token parsing code would hand the comment parser an invalid
token, which might indicate it isn't a comment which is valid or that
the token parser failed to parse the next token to contain the entire
comment which would result in an invalid state for that token and the
program. This change allows the comment code to read a full comment from
the entire text string in the file instead of being restricted to the
output of the token parser allowing for successful comment parsing.

This code still crashes at this moment as it generates multiple root
nodes which the program is next expected to handle. Requesting a review
from Alex Barbur on how to proceed.

* take last root node and optimize comment checking

This enforces that the last root node will be delivered as the primary
root node to the Document. This also optimizes the time to parse a
comment by reducing redudant regex statements using find and match.

* Refactor code for style and fix corner case for empty comments.

* Rename index variables for clarity.

* Refactor to remove a level of indentation.

---------

Co-authored-by: Alexandru Barbur <alex@ctrlc.name>